move tests to __tests__
