+++ __tests__/lib/id.test.ts
@@ -1,5 +1,5 @@
 import { describe, expect, it } from "vitest";
-import { generateId } from "./id";
+import { generateId } from "../../src/lib/id";
 
 describe("generateId", () => {
   it("returns the nanoid output", () => {
