+++ vitest.config.ts
@@ -3,9 +3,8 @@ import { defineConfig } from "vitest/config";
 export default defineConfig({
   test: {
     environment: "node",
-    include: ["src/**/*.{test,spec}.{ts,tsx}"],
+    include: ["__tests__/**/*.{test,spec}.{ts,tsx}"],
     exclude: ["node_modules", "dist"],
     globals: true,
   },
 });
-
